+++ Views/LicenseControl.xaml.cs
@@ -2,23 +2,29 @@ using LicenseManagement.EndUser.Wpf.ViewModels;
 using System;
 using System.Collections.Specialized;
 using System.Configuration;
-using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Media;
 using System.Windows.Media.Imaging;
-using System.Windows.Threading;
 
-namespace LicenseManagement.EndUser.Wpf
+namespace LicenseManagement.EndUser.Wpf.Views
 {
     /// <summary>
-    /// Interaction logic for MainWindow.xaml
+    /// A UserControl for displaying and managing license information.
+    /// Can be embedded directly in any WPF application.
     /// </summary>
-    public partial class MainWindow : Window
+    /// <example>
+    /// XAML usage:
+    /// <code>
+    /// &lt;Window xmlns:lm="clr-namespace:LicenseManagement.EndUser.Wpf.Views;assembly=LicenseManagement.EndUser.Wpf"&gt;
+    ///     &lt;lm:LicenseControl /&gt;
+    /// &lt;/Window&gt;
+    /// </code>
+    /// </example>
+    public partial class LicenseControl : UserControl
     {
-        public MainWindow()
+        public LicenseControl()
         {
             InitializeComponent();
         }
@@ -34,23 +40,22 @@ namespace LicenseManagement.EndUser.Wpf
             get => GetValue(LicenseProperty) as LicenseViewModel;
             set => SetValue(LicenseProperty, value);
         }
+
         public static void LicenseChanged(object sender, DependencyPropertyChangedEventArgs e)
         {
-            var win = sender as MainWindow;
-            win.License = e.NewValue as LicenseViewModel;
-            //win.OnLicenseChanged(e);
+            var control = sender as LicenseControl;
+            control.License = e.NewValue as LicenseViewModel;
         }
 
         public virtual void OnLicenseChanged(DependencyPropertyChangedEventArgs e)
         {
             var newVal = e.NewValue as LicenseViewModel;
-            //newVal.
         }
+
         public static PropertyMetadata Meta
         {
             get
             {
-
                 if (ConfigurationManager.AppSettings.Count != 0)
                 {
                     if (!uint.TryParse(ConfigurationManager.AppSettings.Get("validDays"), out uint valid))
@@ -60,16 +65,18 @@ namespace LicenseManagement.EndUser.Wpf
                     {
                         Expires = DateTime.Now,
                         ValidDays = valid,
-                        //ProductId = ConfigurationManager.AppSettings.Get("productId") ?? "",
                         VendorId = ConfigurationManager.AppSettings.Get("vendorId") ?? "",
                         PublicKey = ConfigurationManager.AppSettings.Get("publicKey") ?? "",
                         ApiKey = ConfigurationManager.AppSettings.Get("ApiKey") ?? "",
                     };
                     var ps = (NameValueCollection)ConfigurationManager.GetSection("Products");
                     defaultValue.Products = new System.Collections.ObjectModel.ObservableCollection<ProductViewModel>();
-                    foreach (var item in ps.AllKeys)
+                    if (ps != null)
                     {
-                        defaultValue.Products.Add(new ProductViewModel { Id = item, Name = ps[item] });
+                        foreach (var item in ps.AllKeys)
+                        {
+                            defaultValue.Products.Add(new ProductViewModel { Id = item, Name = ps[item] });
+                        }
                     }
                     defaultValue.Product = defaultValue.Products.FirstOrDefault();
                     var model = new PropertyMetadata(defaultValue, propertyChangedCallback: LicenseChanged);
@@ -82,39 +89,28 @@ namespace LicenseManagement.EndUser.Wpf
             }
         }
 
-        static MainWindow()
+        static LicenseControl()
         {
             LicenseProperty = DependencyProperty.Register(nameof(License),
                 typeof(LicenseViewModel),
-                ownerType: typeof(MainWindow),
+                ownerType: typeof(LicenseControl),
                 typeMetadata: Meta);
         }
         #endregion
 
-        private void OnProductsSlectionChanged(object sender, System.Windows.Controls.SelectionChangedEventArgs e)
+        private void OnProductsSelectionChanged(object sender, SelectionChangedEventArgs e)
         {
             if (e.AddedItems.Count > 0)
             {
                 var newSelection = e.AddedItems[0] as ProductViewModel;
-                License.CheckLiceneFile(this);
+                License?.CheckLiceneFile(this);
             }
         }
 
-
-        private void thisView_Loaded(object sender, RoutedEventArgs e)
+        private void Control_Loaded(object sender, RoutedEventArgs e)
         {
-
             productsComboBox.SelectedIndex = 0;
-            //var presenter = productsComboBox.Template.FindName("PART_ItemsPresenter", productsComboBox);
-            // Example: Access ComboBoxItem safely
 
-            //var item = productsComboBox.ItemContainerGenerator.ContainerFromIndex(0) as ComboBoxItem;
-            //if (item != null)
-            //{
-            //    Debug.WriteLine("ComboBoxItem is ready: " + item.Content);
-            //}
-
-            //License.RenewLicenseFileAction(this);
             if (Application.ResourceAssembly == null)
             {
                 Application.ResourceAssembly = Assembly.GetExecutingAssembly();
@@ -123,9 +119,6 @@ namespace LicenseManagement.EndUser.Wpf
             {
                 licenseImage.Source = new BitmapImage(new Uri("pack://application:,,,/LicenseManagement.EndUser.Wpf;component/Assets/license.png", UriKind.Absolute));
             }
-            //var rebuilt = productsComboBox.ApplyTemplate();
-
-            // You can also inspect or adjust bindings here
         }
     }
 }